controllers and templates were renamed
