+++ src/wwwroot/app/main.scss
@@ -127,5 +127,9 @@ i.profile-icons {
 .vacancy-description p {
    margin-bottom: 0;
 }
+hr {
+   margin-top: $rem-base/2;
+   margin-bottom: $rem-base;
+}
 
 }
