+++ src/wwwroot/app/views/vacancies/vacancies.search.controller.js
@@ -62,11 +62,11 @@ export default function VacanciesController(
    }
 
    function editVacancy(vacancy) {
-      $state.go('vacancy', {_data: vacancy, vacancyId: vacancy.id});
+      $state.go('vacancyEdit', {_data: vacancy, vacancyId: vacancy.id});
    }
 
    function viewVacancy(vacancy) {
-      $state.go('vacancyProfile', {_data: vacancy, vacancyId: vacancy.id});
+      $state.go('vacancyView', {_data: vacancy, vacancyId: vacancy.id});
    }
 
    function cancel() {
