+++ src/wwwroot/app/views/vacancies/vacancies.search.html
