+++ src/wwwroot/app/views/vacancy.profile/vacancy.view.controller.js
@@ -1,70 +1,71 @@
-const LIST_OF_THESAURUS = ['industries', 'levels', 'locations', 'languages',
-    'departments', 'tags', 'skills', 'typesOfEmployment', 'languageLevels'];
+const LIST_OF_THESAURUS = ['stages', 'industries'];
 import {
    remove
 } from 'lodash';
 
-export default function VacancyController(
+export default function VacancyProfileController(
    $scope,
-   $translate,
    $state,
-   VacancyService,
-   ValidationService,
-   FileUploaderService,
+   $translate,
+   $element,
    ThesaurusService,
    UserService,
+   FileUploaderService,
+   VacancyService,
    UserDialogService,
    HttpService
-) {
+   ) {
    'ngInject';
 
    const vm = $scope;
-
-   /* --- api --- */
-   vm.clear = clear;
-   vm.saveVacancy = saveVacancy;
-   vm.vacancy =  {} ;
-   vm.vacancy.files = $state.params._data ? $state.params._data.files : [];
    vm.thesaurus = [];
    vm.responsibles = [];
+   vm.edit = edit;
+   vm.vacancy =  {};
+   vm.vacancy.files = $state.params._data ? $state.params._data.files : [];
    vm.uploader = createNewUploader();
-   vm.vacancy.requiredSkills = vm.vacancy.requiredSkills || [];
-   vm.vacancy.tags = vm.vacancy.tags || [];
    vm.removeFile = removeFile;
-   vm.errorMessageFromFileUploader = '';
+   vm.saveChanges = saveChanges;
+   vm.changed = changed;
+   vm.isChanged = false;
+   vm.selectStage = selectStage;
+   vm.currentStage = '';
    vm.isFilesUploaded = false;
-   /* === impl === */
 
    function _initCurrentVacancy() {
       if ($state.params._data) {
          vm.vacancy = $state.params._data;
-      } else if ($state.params.vacancyId) {
+      } else {
          VacancyService.getVacancy($state.params.vacancyId).then((vacancy) => {
             vm.vacancy = vacancy;
          });
-      } else {
-         vm.vacancy = {};
       }
    }
-
    _initCurrentVacancy();
+
+   UserService.getUsers().then((users) => {
+      vm.responsibles = users;
+   });
 
    ThesaurusService.getThesaurusTopicsGroup(LIST_OF_THESAURUS).then((data) => vm.thesaurus = data);
 
-   UserService.getUsers().then(users => vm.responsibles = users);
-
    function createNewUploader() {
-      let newUploader = FileUploaderService.getFileUploader({ onCompleteAllCallBack : _vs, maxSize : 2048000 });
+      let newUploader = FileUploaderService.getFileUploader({ onCompleteAllCallBack : saveChanges, maxSize : 2048000 });
       newUploader.onSuccessItem = function onSuccessUpload(item) {
          let response = JSON.parse(item._xhr.response);
          vm.vacancy.files.push(response);
          vm.isFilesUploaded = true;
+         vm.isChanged = false;
       };
       newUploader.onWhenAddingFileFailed = function onAddingFileFailed() {
          vm.errorMessageFromFileUploader = $translate.instant('COMMON.FILE_UPLOADER_ERROR_MESSAGE');
       };
+      newUploader.onAfterAddingAll = function onAfterAddingAl() {
+         vm.isChanged = true;
+      };
       return newUploader;
    }
+
    function removeFile(file) {
       let url = `files/${file.id}`;
       HttpService.remove(url, file).then(() => {
@@ -73,20 +74,24 @@ export default function VacancyController(
       });
    }
 
-   function clear() {
-      $state.go('vacancy', {_data: null, vacancyId: null});
+   function edit() {
+      $state.go('vacancyEdit', {_data: vm.vacancy, vacancyId: vm.vacancy.id});
    }
 
-   function saveVacancy(ev, form) {
-      ev.preventDefault();
-      if (ValidationService.validate(form)) {
-         if (vm.uploader.getNotUploadedItems().length) {
-            vm.uploader.uploadAll();
-         } else {
-            _vs();
-         }
+   function saveChanges() {
+      if (vm.uploader.getNotUploadedItems().length) {
+         vm.uploader.uploadAll();
+      } else {
+         _vs();
       }
-      return false;
+   }
+
+   function changed() {
+      vm.isChanged = true;
+   }
+
+   function selectStage(stageName) {
+      vm.currentStage = stageName;
    }
 
    function _vs() {
@@ -94,6 +99,7 @@ export default function VacancyController(
          vm.vacancy = vacancy;
          vm.vacancy.files = vacancy.files;
          UserDialogService.notification($translate.instant('DIALOG_SERVICE.SUCCESSFUL_SAVING'), 'success');
+         vm.isChanged = false;
       }).catch(() => {
          UserDialogService.notification($translate.instant('DIALOG_SERVICE.ERROR_SAVING'), 'error');
       });
