+++ src/wwwroot/app/views/vacancy.profile/vacancy.view.html
@@ -75,6 +75,7 @@
          <div class="card-divider">
             {{::'CANDIDATES.CANDIDATES' | translate}}
          </div>
+         <hr>
          <div class="card-section">
             <div class="grid-block">
                <div class="small-3 grid-block">
@@ -195,8 +196,19 @@
       </div>
       <div class="main-card vacancy-profile">
          <div class="card-divider">
-            {{::'COMMON.HISTORY' | translate}}
+           <div class="grid-block">
+              <div class="small-6 grid-block">
+              {{::'COMMON.HISTORY' | translate}}
+              </div>
+              <div class="small-3 grid-block">
+                 <a>{{::'COMMON.SHOW_ALL_ACTIONS' | translate}}</a>
+              </div>
+              <div class="small-3 grid-block">
+                 <a>{{::'COMMON.SHOW_COMMENTS_ONLY' | translate}}</a>
+              </div>
+           </div>
          </div>
+         <hr>
          <div class="card-section">
 
          </div>
